refactor: 폴더 구조 변경
- ProductNameValidator, ProductNameConstraint는 Product 도메인에만
  쓰이기 대문에 product 폴더에 두는 게 맞는 것 같음
